Move caching methods out of providers.py into seperate module.
